% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\docType{data}
\name{default_res}
\alias{default_res}
\title{Default shape/rate for prior on r_eff}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
default_res
}
\description{
See vignette("default_priors")
}
\keyword{datasets}
