% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\docType{data}
\name{default_sisc}
\alias{default_sisc}
\title{Default scale for prior on concentration}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
default_sisc
}
\description{
See vignette("default_priors")
}
\keyword{datasets}
