% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.assert_r_realpos}
\alias{.assert_r_realpos}
\title{Unified error message everywhere we assert that R > 0}
\usage{
.assert_r_realpos(r)
}
\description{
Unified error message everywhere we assert that R > 0
}
\keyword{internal}
