% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_nbbp.R
\name{.convert_stan_par}
\alias{.convert_stan_par}
\title{Converts between R,k parameterization and that used internally in stan}
\usage{
.convert_stan_par(params, r_geq_1, to_stan = TRUE)
}
\arguments{
\item{params}{length 2 vector, either R and k or log(R) and log(1/sqrt(k)).}

\item{r_geq_1}{if R must be at least 1 (there are non-extinct, uncensored chains)
stan bounds R at 1, otherwise at 0.}

\item{to_stan}{if TRUE, we take the parameters R and k and put them on stan's unconstrained scale,
if FALSE, we take stan's unconstrained values and put them on the constrained scale.}
}
\value{
vector of transformed parameters
}
\description{
Stan works with log(R) and log(1/sqrt(k)).
}
\keyword{internal}
