% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbbp.R
\name{.dnbbp_subcrit}
\alias{.dnbbp_subcrit}
\title{Optionally-conditioned log-scale PMF of the NBBP for completely-observed finite chain sizes}
\usage{
.dnbbp_subcrit(x, r, k, cond_prob)
}
\arguments{
\item{cond_prob}{the pre-computed conditioning probability to be used in the density;
1.0 if not conditioning on extinction, otherwise Pr(extinct | r, k).}
}
\description{
Arguments as dnbbp except
}
\keyword{internal}
