% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_nbbp.R
\name{.fit_nbbp_homogenous_ml}
\alias{.fit_nbbp_homogenous_ml}
\title{Maximum likelihood fitting function for NBBP}
\usage{
.fit_nbbp_homogenous_ml(
  all_outbreaks,
  censor_geq,
  condition_geq,
  partial_geq,
  partial_probs,
  partial_size_max,
  partial_size_max_error,
  run_reps,
  max_tries,
  seed,
  ...
)
}
\description{
Maximum likelihood fitting function for NBBP
}
\keyword{internal}
