% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_nbbp.R
\name{.get_nonpartial_stan_data}
\alias{.get_nonpartial_stan_data}
\title{Make the part of the stan data input list which pertains to non partially-observed chains}
\usage{
.get_nonpartial_stan_data(partitioned_data)
}
\arguments{
\item{partitioned_data}{output of .partition_data}
}
\description{
This includes completely observed chains, infinite size chains, censored chains,
and accounts for any necessary conditioning on seeing chains of at least a
given size.
}
\keyword{internal}
