% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_nbbp.R
\name{.get_partial_stan_data}
\alias{.get_partial_stan_data}
\title{Make the part of the stan data input list which pertains to partially-observed chains}
\usage{
.get_partial_stan_data(
  partitioned_data,
  partial_size_max,
  partial_size_max_error
)
}
\arguments{
\item{partitioned_data}{output of .partition_data}
}
\description{
Make the part of the stan data input list which pertains to partially-observed chains
}
\keyword{internal}
