% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_partial_ub}
\alias{.get_partial_ub}
\title{Obtains (possibly dynamic) maximum unobserved size of binomially-observed chains}
\usage{
.get_partial_ub(obs_sizes, obs_probs, size_max, error_max)
}
\arguments{
\item{obs_sizes}{the number of cases in all incompletely observed chains}

\item{obs_probs}{the per-case observation probabilities for all chains in \code{obs_sizes}}

\item{size_max}{optional integer override directly specifying the maximum unobserved chain size;
if this is not NA, it will be used regardless of \code{error_max}, though with a warning if it yields
a smaller truncation.}

\item{error_max}{binomial tail probability for dynamic size computation (see details)}
}
\value{
maximum size for summing observation probabilities
}
\description{
This function is intended for internal use only.
}
\details{
If \code{size_max} is an integer, it will be returned. Otherwise, following the math outlined
in the "Implementation details" vignette, an upper bound is chosen such that the error in the marginal
probability of observing a chain of size c is no more than \code{error_max}.
}
\keyword{internal}
