% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbbp.R
\name{.handle_conditioning}
\alias{.handle_conditioning}
\title{Vectorization of extinction and conditioning probabilities for use in dnbbp}
\usage{
.handle_conditioning(x, r, k, condition_on_extinction)
}
\value{
a conditioning probability to pass to .dnbbp_subcrit for each
combination of r, k, and condition_on_extinction
}
\description{
Accounts for vector recycling when any or x, r, and k are not scalars
while maintaining speed when only x is a vector to keep rnbbp efficient.
}
\keyword{internal}
