% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_nbbp.R
\name{.par_boot_nbbp_homogenous}
\alias{.par_boot_nbbp_homogenous}
\title{Parametric bootstrap for NBBP confidence intervals}
\usage{
.par_boot_nbbp_homogenous(
  fit,
  all_outbreaks,
  censor_geq,
  condition_geq,
  ci_width,
  nboot,
  max_tries,
  seed,
  ...
)
}
\description{
Parameters as in fit_nbbp_homogenous_ml, plus fit, the output of .fit_nbbp_homogenous_ml
}
\keyword{internal}
