% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_nbbp.R
\name{.partition_data}
\alias{.partition_data}
\title{Check some errors in and put chain size input data into more useful forms}
\usage{
.partition_data(
  all_outbreaks,
  censor_geq,
  condition_geq,
  partial_geq,
  partial_probs
)
}
\description{
Splits (1) the single vector of all chain sizes up by what likelihood is used,
(2) the single size conditioning vector up by whether the chain is partially
observed or not, and (3) the per-chain observation probabilities into only those
which correspond to actually partially observed chains.
}
\details{
Does NOT account for places where we need to subtract 1 for stan interface.

Arguments as in fit_nbbp_homogenous_bayes
}
\keyword{internal}
