% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_nbbp.R
\name{.prof_nbbp_homogenous_lnl}
\alias{.prof_nbbp_homogenous_lnl}
\title{Profiling the NBBP likelihood for confidence intervals}
\usage{
.prof_nbbp_homogenous_lnl(
  fit,
  all_outbreaks,
  censor_geq,
  condition_geq,
  ci_width,
  ...
)
}
\description{
Parameters as in fit_nbbp_homogenous_ml, plus fit, the output of .fit_nbbp_homogenous_ml
}
\keyword{internal}
