% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.rng_param_recycler}
\alias{.rng_param_recycler}
\title{For vectorizing rnbbp, recycles parameter counts and computes number of sims at each}
\usage{
.rng_param_recycler(n, ...)
}
\value{
list, each item containing n, the number of items to simulate, and the values
of all parameters to be simulated. Note that n may be 0.
}
\description{
For vectorizing rnbbp, recycles parameter counts and computes number of sims at each
}
\keyword{internal}
