% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbbp.R
\name{.simulate_all_offspring}
\alias{.simulate_all_offspring}
\title{Simulate one chain where all offspring counts are known}
\usage{
.simulate_all_offspring(r, k, max_size)
}
\description{
See simulate_all_offspring
}
\keyword{internal}
