% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_nbbp.R
\name{.stan_data_nbbp_homogenous}
\alias{.stan_data_nbbp_homogenous}
\title{Make stan data object for homogenous NBBP model}
\usage{
.stan_data_nbbp_homogenous(
  all_outbreaks,
  censor_geq,
  condition_geq,
  prior,
  likelihood,
  shape_r_eff,
  rate_r_eff,
  sigma_inv_sqrt_dispersion
)
}
\arguments{
\item{prior}{should the prior be included in joint density computations?}

\item{likelihood}{should the likelihood be included in joint density computations?}
}
\description{
Arguments as in fit_nbbp_homogenous_bayes, plus
}
\keyword{internal}
