% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.stop_if_not_mle_enabled}
\alias{.stop_if_not_mle_enabled}
\title{Attempts to ensure MLE is not used without user being aware of risks}
\usage{
.stop_if_not_mle_enabled()
}
\description{
Attempts to ensure MLE is not used without user being aware of risks
}
\keyword{internal}
