% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_nbbp.R
\name{fit_nbbp_all_offspring}
\alias{fit_nbbp_all_offspring}
\title{Fit a negative binomial branching process model where all offspring counts are known}
\usage{
fit_nbbp_all_offspring(
  all_outbreaks_all_offspring,
  censor_geq = rep(NA, length(all_outbreaks)),
  condition_geq = rep(NA, length(all_outbreaks)),
  shape_r_eff = nbbp::default_res,
  rate_r_eff = nbbp::default_res,
  sigma_inv_sqrt_dispersion = nbbp::default_sisd,
  iter = 5000,
  control = list(adapt_delta = 0.9),
  ...
)
}
\arguments{
\item{all_outbreaks_all_offspring}{list (one item per chain) of vectors, each giving the
number of offspring of all individuals in the chain. This must include 0 offspring counts!}

\item{shape_r_eff}{shape parameter of Gamma prior on r_eff.}

\item{rate_r_eff}{rate parameter of Gamma prior on r_eff.}

\item{sigma_inv_sqrt_dispersion}{scale of HalfNormal prior on 1 / sqrt(dispersion).}

\item{iter}{number of iterations for rstan::sampling, default of 5000 intends to be conservative.}

\item{control}{list for rstan::sampling, default attempts to set adapt_delta conservatively.}

\item{...}{further values past to rstan::sampling.}
}
\value{
an rstan stan_fit object
}
\description{
Fit a negative binomial branching process model where all offspring counts are known
}
\details{
When the number of offspring produced by every infection is known, the likelihood is
simply a product of NegativeBinomial(offspring # | R, k) likelihoods. This function
encapsulates a simple stan model which uses the same priors and produces output in
the same form as \link[fit_nbbp_homogenous_bayes]{fit_nbbp_homogenous_bayes}.
}
