% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_nbbp.R
\name{fit_nbbp_homogenous_ml}
\alias{fit_nbbp_homogenous_ml}
\title{Fit a negative binomial branching process model with maximum likelihood inference via rstan}
\usage{
fit_nbbp_homogenous_ml(
  all_outbreaks,
  censor_geq = rep(NA, length(all_outbreaks)),
  condition_geq = rep(NA, length(all_outbreaks)),
  ci_width = 0.95,
  nboot = 1000,
  run_reps = 10,
  max_tries = 50,
  ci_method = "hybrid",
  seed = NA,
  ...
)
}
\arguments{
\item{all_outbreaks}{vector containing the size of each outbreak, including the index case}

\item{censor_geq}{optional, possibly per-chain, censoring, see details.}

\item{condition_geq}{optional, possibly per-chain, conditioning on minimum observed chain size,
see details in \link[nbbp]{fit_nbbp_homogenous_bayes}.}

\item{ci_width}{target width of the 95\% confidence interval (coverage is not guaranteed)}

\item{nboot}{number of parametric bootstrap replicates used for the 95\% CI}

\item{run_reps}{minimum number of random initializations for checking optimization performance,
see details}

\item{max_tries}{maximum number of random initializations when encountering optimization errors,
see details}

\item{ci_method}{adjusts range of r_eff in which likelihood profiling is done instead of
parametric bootstrapping, see details.}

\item{seed}{if not NA, specifies the first seed used in the first rstan::optimizing call spawned
by calling this function. Subsequent calls will modify this seed predictably, ensuring no two
analyses use the same seed. If NA, each seed is the result of calling
sample.int(.Machine$integer.max, 1), as in rstan::optimizing.}

\item{...}{further arguments passed to rstan::optimizing.}
}
\value{
A list the same as \link[rstan]{optimizing} with the following additional components.
$ci: named matrix containing the confidence intervals at the specified width for
r_eff and the dispersion parameter.
$ci_method: character specifying whether the CI was from the "parametric_bootstrap" or
the "likelihood_profile" approach.
$convergence: named matrix containing min and max values for r_eff, the dispersion parameter,
and the log-likelihood
}
\description{
Fit a negative binomial branching process model with maximum likelihood inference via rstan
}
\details{
The likelihood is that used by \link[nbbp]{fit_nbbp_homogenous_bayes}.

Unless the user passes initialization information to rstan, rstan's default of random
initialization is used.
Optimization will be attempted at least \code{run_reps} times, with the reported parameters taken
from the replicate with the highest log-likelihood.
Run-to-run variation is reported as the range of values of the log-likelihood, r_eff, and
dispersion.
If issues are encountered in optimizing, up to \code{max_tries} attempts will be made to obtain
\code{run_reps} successful replicates. Results will be returned even if no replicate ran
without incident. See \link[rstan]{optimizing} for more on the provided \code{return_code}.

Confidence intervals are provided, though in testing the coverage may be relatively different
from the specified width.
The methods applied are not resampling based and thus can provide confidence intervals even if
only index cases have been observed.
By default, confidence intervals are provided by adaptively choosing between intervals provided
by parametric bootstrapping and likelihood profiling (ci_method = "hybrid"), though this can
be changed (ci_method = "boot" for parametric bootstrapping, ci_method = "profile" for
profiling).
The parametric bootstrap has been observed to yield confidence intervals which generally
are more consistent in coverage across parameter values than the method of profiling the
likelihood surface, except near r_eff = 1, where the coverage is quite bad and profiling
the likelihood works better.
}
