% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{measles_us_97}
\alias{measles_us_97}
\alias{measles_canada_98}
\title{Measles chain size data used by Blumberg et al. (2014)}
\format{
Vector of all chain sizes.

\describe{}
}
\source{
Data retrieved from https://doi.org/10.1371/journal.ppat.1004452
}
\usage{
data(measles_us_97)

data(measles_canada_98)
}
\description{
Blumberg et al. (2014) provide a convenient table of measles cases in the US from 1997-1999 and
Canada from 1998-2001. The US data originate from Gay et al. (2004), the Canada data from King
et al. (2004).
}
\examples{
\dontrun{
data(measles_us_97)
data(measles_canada_98)

model <- compile_nbbp_homogenous()
fit_nbbp_homogenous_bayes(
  measles_us_97,
  model
)

fit_nbbp_homogenous_bayes(
  measles_canada_98,
  model
)
}
}
\references{
Blumberg S, Funk S, Pulliam JRC (2014). "Detecting differential transmissibilities that
affect the size of self-Limited outbreaks." PLOS Pathogens 10(10): e1004452.

Gay NJ, De Serres G, Farrington CP, Redd SB, J M (2004). "Assessment of the status of measles
elimination from reported outbreaks: United States, 1997-1999." The Journal of Infectious
Diseases 189 Suppl: S36-S42.

King A, Varughese P, De Serres G, Tipples GA, Waters J, et al. (2004). "Measles elimination in
Canada." The Journal of Infectious Diseases 189 Suppl: S236–42.
}
\keyword{datasets}
