% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mers_pre_june}
\alias{mers_pre_june}
\alias{mers_post_june}
\title{MERS-CoV chain size data used by Blumberg et al. (2014)}
\format{
Vector of all chain sizes.

\describe{}
}
\source{
Data retrieved from https://doi.org/10.1371/journal.ppat.1004452
}
\usage{
data(mers_pre_june)

data(mers_post_june)
}
\description{
Blumberg et al. (2014) provide a convenient table of MERS-CoV cases in the Arabian Peninsula
occurring before August 8, 2013. The dataset is divided into cases occurring before June 1, 2013
and cases occurring after. The data originate from Cauchemez et al. (2014).
}
\examples{
\dontrun{
data(mers_pre_june)
data(mers_post_june)

model <- compile_nbbp_homogenous()
fit_nbbp_homogenous_bayes(
  mers_pre_june,
  model
)

fit_nbbp_homogenous_bayes(
  mers_post_june,
  model
)
}
}
\references{
Blumberg S, Funk S, Pulliam JRC (2014). "Detecting differential transmissibilities that
affect the size of self-Limited outbreaks." PLOS Pathogens 10(10): e1004452.

Cauchemez S, Fraser C, Van Kerkhove MD, Donnelly CA, Riley S, et al. (2014). "Middle
east respiratory syndrome coronavirus: quantification of the extent of the epidemic,
surveillance biases, and transmissibility." The Lancet infectious diseases 14: 50–56.
}
\keyword{datasets}
