% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbbp.R
\name{nb_param_convert}
\alias{nb_param_convert}
\title{Negative binomial parameterization conversion}
\usage{
nb_param_convert(size = NULL, prob = NULL, mu = NULL)
}
\arguments{
\item{size}{either the number of samples (probability-size
parameterization) or the dispersion parameter (mean-dispersion
parameterization)}

\item{prob}{probability of success in each trial}

\item{mu}{mean}
}
\value{
named vector providing alternative parameterization for
use with stats::rnbinom and related functions.
}
\description{
Negative binomial parameterization conversion
}
\details{
Converts between the mean-dispersion parameterization and the
probability-size parameterization of the Negative Binomial.
Supply either prob and size to get mu (mean) and size (dispersion), or
mu and size to get prob and size.
}
