% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbbp-package.R
\docType{package}
\name{nbbp-package}
\alias{nbbp}
\alias{nbbp-package}
\title{nbbp: Stuttering chains simulations}
\description{
Inference and simulation for stuttering chains under a negative binomial branching process.
}
\author{
\strong{Maintainer}: Andrew Magee \email{rzg0@cdc.gov}

Authors:
\itemize{
  \item Scott Olesen \email{ulp7@cdc.gov}
}

}
\keyword{internal}
