% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbbp.R
\name{nbbp_small_r}
\alias{nbbp_small_r}
\title{NBBP reproduction number symmetry}
\usage{
nbbp_small_r(r, k)
}
\arguments{
\item{r}{effective reproduction number, must be > 1}

\item{k}{dispersion parameter}
}
\value{
the corresponding subcritical R < 1
}
\description{
NBBP reproduction number symmetry
}
\details{
Waxman and Nouvellet(https://doi.org/10.1016/j.jtbi.2019.01.033) show that,
when conditioning on extinction, any supercritical R > 1 can be mapped to
a subcritical R < 1 (keeping k constant) which produces identical distributions
on the final chain size.
}
