% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pneumonic_plague}
\alias{pneumonic_plague}
\title{Chain size data for pneumonic plague}
\format{
Vector of all chain sizes.
}
\source{
https://doi.org/10.1016/j.jtbi.2011.10.039
}
\usage{
data(pneumonic_plague)
}
\description{
Nishiura et al. (2012) report 19 outbreaks with more than one case of pneumonic plague since
1900. Given the collection process, analysis must condition on seeing chains of size >= 2.
Further, the chain of size 5009 should either be treated as censored (at size 5009) or
re-coded to Inf.
}
\examples{
\dontrun{
data(pneumonic_plague)

# Fit with censoring
model <- compile_nbbp_homogenous()
fit_nbbp_homogenous_bayes(
  pneumonic_plague,
  model,
  condition_geq = rep(2, 19),
  censor_geq = c(rep(NA, 3), 5009, rep(NA, 15))
)

# Fit with re-coding
pneumonic_plague[pneumonic_plague == 5009] <- Inf
fit_nbbp_homogenous_bayes(
  pneumonic_plague,
  model,
  condition_geq = rep(2, 19)
)
}
}
\references{
Nishiura H, Yan P, Sleeman CK, Mode CJ (2012). "Estimating the transmission potential of
super-critical processes based on the final size distribution of minor outbreaks."
Journal of Theoretical Biology 294: 48–55.
}
\keyword{datasets}
