% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prior_predictive}
\alias{prior_predictive}
\title{Prior and prior predictive distributions}
\format{
data.frame with 10,000 samples from the prior and prior predictive distributions.
\describe{
\item{r_eff}{The effective reproduction number R.}
\item{dispersion}{The dispersion parameter k.}
\item{offspring_count}{One sample from offspring distribution given by R, k.}
\item{chain_size}{One sample from chain size distribution given by R, k.}
}
}
\source{
Generated in R using nbbp.
}
\usage{
data(prior_predictive)
}
\description{
Samples from the default joint prior on R and k, as well as the prior predictive distribution
on Pr(0 offspring | R, k) and Pr(chain size | R, k).
}
\examples{
\dontrun{
data(prior_predictive)
hist(prior_predictive$chain_size)
}
}
\keyword{datasets}
