% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sbc_quants}
\alias{sbc_quants}
\title{High-level validation results}
\format{
data.frame with nominal quantiles and observed coverage.
\describe{
\item{quantile}{The quantile for which the CI was computed.}
\item{r_covered}{The proportion of true values of R covered at the quantile.}
\item{r_covered}{The proportion of true values of k covered at the quantile.}
}
}
\source{
Generated in R using nbbp.
}
\usage{
data(sbc_quants)
}
\description{
Coverage of quantiles from 1,000 analyses of datasets simulated under the prior predictive.
}
\details{
When datasets are simulated from the prior predictive, frequentist coverage
properties of the posterior quantiles should hold. This dataframe contains the results
from analyzing 1,000 datasets simulated from the prior predictive (under default priors)
of 25 observations each. Thus, a plot of quantiles and coverage at those quantiles
should fall near the 1:1 line.

Only analyses meeting convergence standards (min(n_eff) > 1000, max(Rhat) < 1.005)
are included in those used to compute coverage.
}
\examples{
\dontrun{
data(sbc_quants)
plot(sbc_quants$quantile, sbc_quants$r_covered)
abline(a = 0, b = 1)
}
}
\keyword{datasets}
