% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_based_testing}
\alias{sim_based_testing}
\title{High-level performance results in simulations}
\format{
data.frame with true values and estimates.
\describe{
\item{index}{Simulation index for replicate simulations with shared (R, k, n_chains).}
\item{n_chains}{The number of chains in the simulated dataset.}
\item{r_true}{The true simulating value of R.}
\item{k_true}{The true simulating value of k.}
\item{r_point}{The posterior median or maximum likelihood value of R.}
\item{k_point}{The posterior median or maximum likelihood value of k.}
\item{r_low}{The lower bound of a 95\% credible of confidence interval for R.}
\item{r_high}{The upper bound of a 95\% credible of confidence interval for R.}
\item{k_low}{The lower bound of a 95\% credible of confidence interval for k.}
\item{k_high}{The upper bound of a 95\% credible of confidence interval for k.}
\item{estimator}{Type of analysis, bayes (Bayesian) or maxlik (maximum likelihood).}
}
}
\source{
Generated in R using nbbp.
}
\usage{
data(sim_based_testing)
}
\description{
Analyses of a grid (over R, k, and number of chains) of datasets.
Point (posterior median, maximum likelihood) and interval (95\% credible interval,
95\% confidence interval) estimates from both Bayesian and maximum likelihood analyses.
}
\examples{
\dontrun{
data(sbc_quants)
plot(sim_based_testing$r_true, sim_based_testing$r_point)
abline(a = 0, b = 1)
}
}
\keyword{datasets}
