% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbbp.R
\name{simulate_all_offspring}
\alias{simulate_all_offspring}
\title{Simulate chains where all offspring counts are known}
\usage{
simulate_all_offspring(n, r, k, max_size = 1000)
}
\arguments{
\item{n}{number of samples to draw}

\item{r}{effective reproduction number}

\item{k}{dispersion parameter: when <1, overdispersed}

\item{max_size}{simulation will be terminated if more than this many individuals
are infected in any one chain.}
}
\description{
Simulate chains where all offspring counts are known
}
\details{
In this model, every individual infects a Negative-Binomially-distributed
number of additional individuals.
}
