% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{table_add_1d}
\alias{table_add_1d}
\title{Addition of two 1-D base::table objects}
\usage{
table_add_1d(t1, t2, keep_zeros = FALSE)
}
\arguments{
\item{t1}{a base::table object}

\item{t2}{a second base::table object}

\item{keep_zeros}{whether we should retain in the result values with total count 0 (possible if
negating one of the added tables).}
}
\value{
a base::table where the count of every label is the sum of the counts in t1 and t2
}
\description{
Addition of two 1-D base::table objects
}
\examples{
table_add_1d(table(1:10), table(5:15))
}
