% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{borealpox}
\alias{borealpox}
\title{Chain size data for borealpox}
\format{
Vector of all chain sizes.
}
\source{
https://en.wikipedia.org/wiki/Borealpox_virus
}
\usage{
data(borealpox)
}
\description{
Per Wikipedia (accessed September 2024), there have been 7 reported single cases of borealpox.
}
\examples{
\dontrun{
data(borealpox)
model <- compile_nbbp_homogenous()
fit_nbbp_homogenous_bayes(borealpox, model)
}
}
\keyword{datasets}
