% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_nbbp.R
\name{compute_likelihood_surface}
\alias{compute_likelihood_surface}
\title{Computes NBBP likelihood surface for given data at given grid}
\usage{
compute_likelihood_surface(
  all_outbreaks,
  r_grid,
  k_grid,
  censor_geq = rep(NA, length(all_outbreaks)),
  condition_geq = rep(NA, length(all_outbreaks))
)
}
\arguments{
\item{all_outbreaks}{vector containing the size of each outbreak, including the index case}

\item{r_grid}{vector of values of R at which the likelihood is to be evaluated.}

\item{k_grid}{vector of values of k at which the likelihood is to be evaluated.}

\item{censor_geq}{optional, possibly per-chain, censoring, see details.}

\item{condition_geq}{optional, possibly per-chain, conditioning on minimum observed chain size,
see details in \link[nbbp]{fit_nbbp_homogenous_bayes}.}
}
\value{
a long-form tibble of R, k, log-likelihood
}
\description{
Computes NBBP likelihood surface for given data at given grid
}
