% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbbp.R
\name{dnbbp}
\alias{dnbbp}
\alias{pnbbp}
\alias{rnbbp}
\title{Final outbreak size for negative binomial branching process}
\usage{
dnbbp(x, r, k, condition_on_extinction = FALSE)

pnbbp(q, r, k)

rnbbp(n, r, k, condition_on_extinction = FALSE, max_size = 1e+06)
}
\arguments{
\item{x}{final outbreak size, including the index case}

\item{r}{effective reproduction number}

\item{k}{dispersion parameter: when <1, overdispersed}

\item{condition_on_extinction}{logical, should we condition_on_extinction the process on
extinction (TRUE) or not (FALSE)}

\item{q}{as \code{x}}

\item{n}{number of samples to draw}

\item{max_size}{when drawing samples, the pmf for non-infinite chains is truncated to
this size determined. Infinite chains are still possible (and return Inf).}
}
\description{
Final outbreak size for negative binomial branching process
}
\details{
In this model, every individual infects a Negative-Binomially-distributed
number of additional individuals.
The distribution is on the final number of infected individuals, including
the index case.

Blumberg et al 2014 (10.1093/aje/kwu068) equation 1, calls the terms in
this PMF r(j), the probability that a transmission chain will
have true size j.

When R >= 1.0, the process may not go extinct. In these cases, the outcome
is dependent on the choice of \code{condition_on_extinction}.

When R >= 1 and condition_on_extinction == FALSE (the default),
chains are allowed to become infinitely large.
In this case, \code{rnbbp} will record this size as Inf.

When R >= 1 and condition_on_extinction == TRUE, the model is conditioned on all chains going
extinct. No infinitely-large chains are allowed.

Nishiura et al 2012 (https://doi.org/10.1016/j.jtbi.2011.10.039) discuss the
R > 1 case when conditioning on extinction.
}
