% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_nbbp.R
\name{fit_nbbp_homogenous_bayes}
\alias{fit_nbbp_homogenous_bayes}
\title{Fit a negative binomial branching process model with Bayesian inference via rstan}
\usage{
fit_nbbp_homogenous_bayes(
  all_outbreaks,
  censor_geq = rep(NA, length(all_outbreaks)),
  condition_geq = rep(NA, length(all_outbreaks)),
  mu_r_eff = 0,
  sigma_r_eff = 0.421404,
  sigma_inv_sqrt_dispersion = 1.482602,
  iter = 5000,
  ...
)
}
\arguments{
\item{all_outbreaks}{vector containing the size of each outbreak, including the index case}

\item{censor_geq}{optional, possibly per-chain, censoring, see details.}

\item{condition_geq}{optional, possibly per-chain, conditioning on minimum observed chain size,
see details.}

\item{mu_r_eff}{location parameter of Lognormal prior on r_eff (prior mean(log(r_eff))).}

\item{sigma_r_eff}{scale parameter of Lognormal prior on r_eff (prior sd(log(r_eff))).}

\item{sigma_inv_sqrt_dispersion}{scale of HalfNormal prior on 1 / sqrt(dispersion).}

\item{iter}{number of iterations for rstan::sampling.}

\item{...}{further values past to rstan::sampling.}
}
\value{
an rstan stan_fit object
}
\description{
Fit a negative binomial branching process model with Bayesian inference via rstan
}
\details{
The likelihood is as described in \link[nbbp]{dnbbp}, without conditioning on extinction.

In particular, we fit a Bayesian model in rstan, assuming all cases are observed.
The variables recorded in the stan fit object are (ignoring stan's ordering):
\enumerate{
\item \code{r_eff}, the effective reproduction number R;
\item \code{dispersion}, k, the parameter controlling (over)dispersion;
\item \code{inv_sqrt_dispersion}, 1 / sqrt(k);
\item \code{exn_prob} the probability that the chain goes extinct for this R and k pair;
\item \code{p_0} the probability that one individual has 0 offspring for this R and k pair.
}

It is possible to condition the likelihood on only observing chains of at least a certain
size. This may be done on a per-chain basis, if sampling processes vary. If the ith chain
should be conditioned on seeing a chain of at least size m, set \code{condition_geq[i] = m}.
NA for no conditioning (chains can be any size greater than or equal to 1).

Right-censoring is permitted on a per-chain basis. That is, if there is some size above which
a chain can be said not to go extinct on its own, we can treat this as an observation of
at least that size. If the ith chain size should be treated as censored at size m, set
\code{censor_geq[i] = m}. The value of \code{all_outbreaks[i]} is irrelevant. NA for no censoring.

Conditioning and/or right-censoring reveal that there are small numerical instabilities in the
PMF which can make it sum to values ever so slightly larger than 1. This is handled by
rescaling the CDF such that it does not exceed 1 in the range it needs to be evaluated.

We let r_eff ~ LogNormal(mu = mu_r_eff, sigma = sigma_r_eff)
(parameterized in terms of log-mean and log-sd as in stan and R),
with the default being Lognormal(mu = 0.0, sigma = 0.421404).
The default has a median R of 1, and a 90\% prior CI of spanning from 1/2 to 2.

We let 1/sqrt(dispersion) ~ HalfNormal(sigma = sigma_inv_sqrt_dispersion),
with the default being HalfNormal(sigma = 1.482602).
The default puts 50\% of the prior mass in the highly-overdispersed
region (dispersion parameter < 1, more dispersed than a Geometric distribution) and 50\%
in the less overdispersed region (dispersion parameter > 1, less dispersed than a Geometric
distribution, with the Poisson limit at infinity).
}
