% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbbp.R
\name{nbbp_stats}
\alias{nbbp_stats}
\title{NBBP conditional mean and variance}
\usage{
nbbp_stats(r, k)
}
\arguments{
\item{r}{effective reproduction number}

\item{k}{dispersion parameter}
}
\value{
named vector providing the (conditioned on there being a finite
chain size, i.e. on extinction) mean and variance of the final size.
}
\description{
Uses equations for mean and variance of chain size from
Waxman and Nouvellet(https://doi.org/10.1016/j.jtbi.2019.01.033)
}
